+++ tests/testthat/test-check_gcae_input_filenames.R
@@ -1,0 +1,8 @@
+test_that("use", {
+  expect_equal(1 + 1, 2) # Prevents testthat warning for empty test
+  if (!plinkr::is_on_ci()) return()
+  if (!plinkr::is_plink_installed()) return()
+
+  gcae_input_filenames <- create_gcae_input_files_1()
+  check_gcae_input_filenames(gcae_input_filenames)
+})